Update stylesheet.css
changed the grid layout so that I can set the column and row templates. It essentially ensures that second-card and third-card will both take up 1fr. 1fr = "a fraction". So in this case they both take up 1 fraction, sharing available space of the grid-parent, which is of course controlled by your max-width.